adding asset and changing cell
