+++ src/xcode/ENA/ENA/Resources/Assets/Assets.xcassets/Common/Icons_admission_state.imageset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "images" : [
+    {
+      "filename" : "icon_admission_state.pdf",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
