+++ src/xcode/ENA/ENA/Resources/Assets/Assets.xcassets/Common/Icons_admission_state.imageset/icon_admission_state.pdf
@@ -1,0 +1,184 @@
+%PDF-1.7
+
+1 0 obj
+  << /Length 2 0 R >>
+stream
+1.161621 0 0.087402 -0.192383 1.074219 1.181152 d1
+
+endstream
+endobj
+
+2 0 obj
+  51
+endobj
+
+3 0 obj
+  [ 1.161621 ]
+endobj
+
+4 0 obj
+  << /Length 5 0 R >>
+stream
+/CIDInit /ProcSet findresource begin
+12 dict begin
+begincmap
+/CIDSystemInfo
+<< /Registry (FigmaPDF)
+   /Ordering (FigmaPDF)
+   /Supplement 0
+>> def
+/CMapName /A-B-C def
+/CMapType 2 def
+1 begincodespacerange
+<00> <FF>
+endcodespacerange
+1 beginbfchar
+<00> <DBC0DFAB>
+endbfchar
+endcmap
+CMapName currentdict /CMap defineresource pop
+end
+end
+endstream
+endobj
+
+5 0 obj
+  336
+endobj
+
+6 0 obj
+  << /Subtype /Type3
+     /CharProcs << /C0 1 0 R >>
+     /Encoding << /Type /Encoding
+                  /Differences [ 0 /C0 ]
+               >>
+     /Widths 3 0 R
+     /FontBBox [ 0.000000 0.000000 0.000000 0.000000 ]
+     /FontMatrix [ 1.000000 0.000000 0.000000 1.000000 0.000000 0.000000 ]
+     /Type /Font
+     /ToUnicode 4 0 R
+     /FirstChar 0
+     /LastChar 0
+     /Resources << >>
+  >>
+endobj
+
+7 0 obj
+  << /Font << /F1 6 0 R >> >>
+endobj
+
+8 0 obj
+  << /Length 9 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 0.000000 cm
+0.000000 0.498039 0.678431 scn
+16.000000 0.000011 m
+24.773207 0.000011 32.000000 7.226805 32.000000 16.000010 c
+32.000000 24.773216 24.757895 32.015320 15.984690 32.015320 c
+7.211484 32.015320 0.000000 24.773216 0.000000 16.000010 c
+0.000000 7.226805 7.226795 0.000011 16.000000 0.000011 c
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 5.000000 8.327820 cm
+1.000000 1.000000 1.000000 scn
+0.000000 2.687500 m
+h
+6.987305 15.422852 m
+6.987305 13.823242 8.059570 12.469727 9.536133 12.065430 c
+9.536133 6.528320 l
+9.536133 4.137695 10.072266 2.661133 10.458984 2.661133 c
+10.836914 2.661133 11.364258 4.128906 11.364258 6.528320 c
+11.364258 12.065430 l
+12.849609 12.460938 13.930664 13.823242 13.930664 15.422852 c
+13.930664 17.338867 12.383789 18.912109 10.458984 18.912109 c
+8.525391 18.912109 6.987305 17.338867 6.987305 15.422852 c
+h
+9.474609 15.229492 m
+8.841797 15.229492 8.279297 15.783203 8.279297 16.424805 c
+8.279297 17.066406 8.841797 17.611328 9.474609 17.611328 c
+10.125000 17.611328 10.661133 17.066406 10.661133 16.424805 c
+10.661133 15.783203 10.125000 15.229492 9.474609 15.229492 c
+h
+10.450195 -0.775391 m
+16.022461 -0.775391 19.335938 1.096680 19.335938 3.241211 c
+19.335938 5.605469 15.706055 7.117188 12.858398 7.231445 c
+12.858398 5.684570 l
+14.791992 5.605469 17.340820 4.691406 17.340820 3.416992 c
+17.340820 1.905273 14.440430 0.859375 10.450195 0.859375 c
+6.459961 0.859375 3.568359 1.896484 3.568359 3.416992 c
+3.568359 4.691406 6.117188 5.605469 8.041992 5.684570 c
+8.041992 7.231445 l
+5.203125 7.117188 1.573242 5.605469 1.573242 3.241211 c
+1.573242 1.096680 4.886719 -0.775391 10.450195 -0.775391 c
+h
+f
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 5.000000 8.327820 cm
+BT
+18.000000 0.000000 0.000000 18.000000 0.000000 2.687500 Tm
+/F1 1.000000 Tf
+[ (\000) ] TJ
+ET
+Q
+
+endstream
+endobj
+
+9 0 obj
+  1879
+endobj
+
+10 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 32.000000 32.015320 ]
+     /Resources 7 0 R
+     /Contents 8 0 R
+     /Parent 11 0 R
+  >>
+endobj
+
+11 0 obj
+  << /Kids [ 10 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+12 0 obj
+  << /Pages 11 0 R
+     /Type /Catalog
+  >>
+endobj
+
+xref
+0 13
+0000000000 65535 f
+0000000010 00000 n
+0000000117 00000 n
+0000000138 00000 n
+0000000169 00000 n
+0000000561 00000 n
+0000000583 00000 n
+0000000995 00000 n
+0000001041 00000 n
+0000002976 00000 n
+0000002999 00000 n
+0000003174 00000 n
+0000003250 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 12 0 R
+   /Size 13
+>>
+startxref
+3311
+%%EOF
