+++ src/xcode/ENA/ENA/Source/Scenes/HealthCertificates/Overview/Cells/ChangeAdmissionScenario/ChangeAdmissionScenarionCellModel.swift
@@ -1,0 +1,19 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import UIKit
+import OpenCombine
+
+class changeAdmissionScenarionCellModel: AddButtonAsTableViewCelling {
+	
+	// MARK: - Internal
+
+	// to.do should be dynamic
+	let text: String = "Regeln des Bundes"
+
+	var iconImagePublisher = CurrentValueSubject<UIImage?, Never>(UIImage(named: "Icons_Admission_State"))
+	var textColorPublisher = CurrentValueSubject<UIColor, Never>(.enaColor(for: .textPrimary1))
+	var accessibilityTraitsPublisher = CurrentValueSubject<UIAccessibilityTraits, Never>([.button])
+
+}
